Split assistant card renderer tests
